Bug fix: we should rename every definition.
Even the first def might take it out of SSA because there might be a read preceding it.
tests/if_else.c is an example.

This creates a few more renames than required, but is correct.
Now compile.sh compiles all examples in tests. Still need to check that
the compilation output is correct, but this is progress for now.