Link against clang libraries on the system (same location as LLVM libs)
We need to specify each library twice because there are circular deps!
https://github.com/eliben/llvm-clang-samples/blob/master/Makefile#L71

We can't use start-group, end-group as Eli's code does because they
don't play well with autotools.
--> Adding them to LDADD isn't possible because linker options belong
in LDFLAGS.
--> Adding them to LDFLAGS doesn't work either because of the way
autotools treats LDFLAGS
http://lists.gnu.org/archive/html/automake/2006-04/msg00096.html

The only workaround is to list everything twice:
http://lists.gnu.org/archive/html/automake/2006-04/msg00098.html

This increases build time by a second or so, which is fine. It's also
a bit fragile because if you have libtool enabled, it replaces multiple
occurences of libraries with one occurence!

So much for autotools :)